Move manager args to config/manager
The auth proxy patch used to add the address for the metrics server,
which meant that the patch used to have to include _all_ the args
(kustomize can't modify a single item in a list, it can only replace the
entire list). However, it doesn't hurt us to pass the metrics address in
the base config for the manager (we're passing in the default value
anyway) and it makes the configs much nicer to _only_ have the auth
proxy config in the auth proxy patch.

Also added a variety of comments.

Tested: verified that other than reordering a few lines, the generated
hnc-manager.yaml manifest file is identical before and after this
change.