Use CRD validation to enforce singletons
Add CRD validation kustomization patchs to HierarchyConfiguration and
HNCConfiguration CRDs to validate the metadata.name to be the singleton.

Remove the "CritSingletonNameInvalid" condition and remove the
validating and condition setting code in the HNCConfig reconciler.
HNCConfig reconiler only reconciles the 'config' singleton. The
validation here is only to set the condition so it's safe to remove.

Remove the validating webhook on HNCConfig resource name too. There's no
need to keep the validating webhook as a second layer of protection
because the server-side validation is more reliable than webhooks since
webhooks can be disabled or bypassed if not running.

This change is also backward compatible since the old bad HNCConfig
resource will still have the obsolete condition on it. HNCConfig
reconciler won't touch it.

Tested on kind and GKE 1.14 cluster. New HC or HNCConfig with bad
names cannot be created with the server-side validation. An old
HNCConfig with bad name created before introducing the CRD validation
when the webhook was also disabled can still be deleted now. There was
still an obsolete "CritSingletonNameInvalid" condition on it.