Allow subnamespace deletion if parent/anchor is missing
Since we allow changes to the subnamespaceOf annotation on subnamespace
as an escape hatch and there's a scenario of recovering from CRD
deletion, this commit makes it easier to delete a subnamespace, but only
when the parent or the anchor is missing. If the anchor is ok (state),
we still don't allow direct deletion of a subnamespace and only allows
deletion from the anchor.

Tested with new unit test cases and manually before and after.