Restore webhook in e2e test from the current manifest
Remove the `kubectl apply -f manifests/hnc-manager.yaml` in the e2e
test as a way to restore webhook. Instead, output the webhook manifest
into a file and later apply this file to restore the webhook. This makes
sure the webhook is the one used in the current deployment when the e2e
test is run.

Tested by running the script and the generated yaml file is deleted
after each test.