Add conversion webhook to HNCConfig for sync mode
This is the initial commit to introduce CRD conversion webhook in HNC.
The conversion of sync mode in HNCConfig is minor so it's a good
starting point.

Set v1alpha1 as the conversion hub and v1alpha2 as a spoke. Add
ConvertFrom() to convert CRD spec from v1alpha1 to v1alpha2. Omit the
ConvertTo() because we only need one-way convertion.

Add conversion strategy and webhookClientConfig in a kustomize patch.
Update the internal cert reconciler to inject cert to the conversion
webhook too. (The cer-manager option is not available for this current
conversion webhook and the caninjectcert patch is not included either)

Tested by setting v1alpha2 as the only served version and storage
version. Got expected spec and empty status. Unable to update the
resource yet with the current global enum schema validation. Tested
manually to set different enum schema for each version and the update
worked. The changes to switch from global schema validation to separate
ones for each version are not included in this PR but will be in the
next PRs.