Make it easier to uninstall HNC
If you have any subnamespaces, simply saying `kubectl delete -f
manifests/hnc-manager.yaml` will not work, since the deployment will be
deleted before the subnamespace finalizers are removed. This change
addes a new `make undeploy` target that removes the CRDs _first_, and
only later deletes the rest of HNC. It's a bit hacky and produces a few
spurious error messages (since we try to delete the CRDs twice), but it
works.

Also removes the unused `deploy-cm` target that used to install
cert-manager, but this is no longer a dependency of HNC.

Tested: tested locally with and without subnamespaces; this robustly
deletes HNC.