Change HNC image name to hnc-manager
The current HNC image name is hnc/controller, which is kind of silly
because the "c" in "HNC" already stands for "controller," and also
because some image registries (e.g. Docker) don't allow paths within
image names. This change modifies the default (and official) image name
to be hnc-manager, which aligns with the existing manifest file name of
hnc-manager.yaml. The change should be invisible to almost all users.

Note that if you modify the default via HNC_IMG while calling `make`,
the manifest file will also change. I think this is a feature, not a bug
- the manifest file with a given name will always refer to the image of
the same name.

Tested: deployed locally