Use discovery to validate types in hncconfig
Currently, the hncconfig validator validate types by making a Get() call
using the controller-runtime client to see if the type exists from the
error types returned. This is indirect and error-prone when the error
types returned fron controller-runtime change.

This commit uses the client-go discovery client to look for a matching
type.

Tested manually that the cannot set-type issue is fixed and requests
with incorrect GVK are still denied.