Fix the flaky test for #716
Change the mustNotRun() function to use Eventually() so that we get
stable test result. Add sleep time after create subnamespace so that the
anchor is ready before test. Change the default wait time in mustRun()
and mustNotRun() to 5 seconds because deleting a subnamespace takes a
while.

Tested: make e2e-test