Fix flaky test
From the fail logs on issue #871, the flaky tests all happen after a
HNCConfigue update, but the supposed follwing updates could not be
detected. We give these lines more time in Eventually() so that
HNCConfigue update gets more time to propogate. However, I was not able
to reproduce the flaky test results when running locally, so I cannot
confirm that they are fully fixed by this pr.

Tested: make test