Fix the Rolebinding update logic
RoleBindings can't have their Roles changed after they're created. If an RB was quickly delete and re-created in an ancestor namespace - fast enough that by the time that HNC notices, the new RB exists - HNC could see it as an update (not a delete + create) and attempt to update the RBs in all descendant namespaces, and this will fail. In order to handle this case, we try to delete and re-create the rolebinding instead.

Tested: make test-e2e