Fix some flaky testes
Extend the timeout limit for some operations to make them more stable.
Checks that the previous operation finishes before proceeding to the
next one, otherwise some of them will be flaky. Warns the user that the
network policy has to be enabled for the test to pass.

Tested: make test-e2e