Update hc validator to check object conflicts
Add an additional rule to make sure there's no potential object
overwriting in hierarchy changes. Format the webhook message to:
"
Cannot update hierarchy because it would overwrite the following object(s):
  * Namespace "team-b": my-creds (/v1, Kind=Secret)
  * Namespace "team-b": my-creds2 (/v1, Kind=Secret)
To fix this, please rename or remove the conflicting objects first.
"

Format the object webhook message to:
"
Cannot create "my-creds2" (/v1, Kind=Secret) in namespace "acme-org" because it would overwrite objects in the following descendant namespace(s):
  * team-a
  * team-b
To fix this, choose a different name for the object, or remove the conflicting objects from the above namespaces.
"

Tested manually on GKE cluster and by 'make test' with new test
cases.