Fix make target kind-deploy for empty PROJECT_ID
Reference issue #1084

Make target kind-deploy was failing due to empty PROJECT_ID if gcloud is not
installed. Setting REGISTRY should not be mandatory for kind
deployments, so skipping the REGISTRY field if PROJECT_ID is empty.

Tested: Running make target kind-deploy fails without this fix and runs
after this change. Covered the testing of scenarios where PROJECT_ID is
set or empty. The deployment goes through for kind version 0.8.1.

Ignore HNC_REGISTRY if PROJECT_ID is not present