Build Krew archive
Unlike in 0.6+, this change doesn't attempt to build the manifest as a
part of Cloud Build. That can be done by hand after the release; too
much has changed in the makefiles between the two versions. So this just
focuses on creating the archive.

Tested: this is virtually impossible to test given the state of the
build process in v0.5 (it's much better in v0.6). I'll revert the
changes to cloudbuild.yaml if I can't build a release candidate, and the
changes to the hack/krew*.yaml files don't matter to anyone.