Refactor forest `sourceObjects` and update object reconciler
Refactor the `forest.namespace.sourceObjects` field to not only store
source objects with `Propagate` mode, but all the source objects as long
as there once is an object reconciler created for the type.

This changed field will provide data for handling source object name
conflicts even for types changing mode from others to `Propagate`. In
addition, since we will implement HNC exceptions soon, we will not simply
rely on whether a source object exists in this field, but will get all
the info on the fly from the instance itself to decide whether to
propagate or enqueue objects.

Update object reonciler dependencies on the old `originalObjects` field
with only source objects with `Propagate` mode. Check if the object
`shouldPropagate()` on the fly. Remove the check of whether the source
object is updated to ensure a mode change from `Remove` to `Propagate`
still enqueues an unchanged source's descendants for them to be created.
Note this will increase the number of enqueued objects but it should be
fine since we only write if there's a diff and read is fast from cache.

Update object reconciler to still not deleting a descendant object with
the same name as in an ancestor for types with 'Remove' mode. Add a test
case for that. The test failed before and passed after the change.

Consider there's an object overwriting conflict only if the conflicting
source in the ancestor can propagate, e.g. with no finalizers. Add a new
test case.

Tested by `make test`.