Add exception when using 'select'
Add `labels` field in namespace struct within Forest. When propagating
an object, we first check if a `select` annotation exists. If this
annotation exists and doesn't match the current namespace, this object
will not be propagated.

Modify the object_test.go to use `Role` instead of `Secret` in
propagating tests.

Tested: make test