Ensure controller-gen is rebuilt properly
The Makefile wasn't correctly regenerating controller-gen when the
version in go.mod (and hence vendor/) was updated. This led to me
mistakenly submitting #1124. This change _always_ rebuilds
controller-gen from scratch to ensure that leftover tool binaries won't
cause more problems.

Tested: before this change, ran `make manifests` and observed that the
version of controller-gen (as seen by `--version`) didn't match the
version in go.mod. After this change, observed that the reported version
now matched that in go.mod and the CRD YAMLs were rebuilt. Installed
HNC and verified that it started up ok.