Set crd:trivialVersions=false for easier API upgrade
We were going to set crd:trivialVersions=false for per-version
validation in 0.6. However, we cannot patch CRD if the one in v0.5 is
still using top-level schema. We would get "Forbidden: top-level and
per-version schemas are mutually exclusive" error when upgrading HNC,
specifically CRDs.

Tested by API upgrading from this to the v0.6 change in a GKE cluster.
It passed after this change and failed with the above error before. Also
tested by 'make test' to ensure it doesn't break v0.5.