Improve anchor deletion
This change fixes todos in the cherrypick for #1150 (see issue #1149).
It simplifies and restructures a lot of the logic to make it easier to
follow while looking at less data (e.g. a lot more focus on
anchor.Status.State). It also adds a lots more documentation.

Tested: all e2e tests pass on GKE 1.18 when combined with fixes to the
e2e tests (PRs #1160, #1162, #1163 and #1164).