Delete bad anchors and fix e2e tests
While running the e2e tests, the CleanupNamespace command was hanging
because we weren't allowing deletion of anchors in the Conflict state;
we were just checking whether the would-be subnamespace had the correct
parent, but not whether the subnamespace anchor was actually in the 'ok'
state. This was resulting in HNC blocking the deletion of a bad anchor,
which we should always allow (this should have been done as a part of
PR #1151).

I also made three minor fixes to the e2e tests themselves:
1. Turn propagation on for new types with the --force option
2. Skip the network policy demo if we detect that network policies
   aren't enabled on the cluster.
3. Delete the webhook configs when HNC might be in a bad state so that
   they don't try to block cleanup of namespaces.

Tested: the e2e tests hang repeatedly and all the demo tests fail
without this fix; with this fix, they pass.