Update HNCConfig API with GR for restmapper lookup
In v1alpha2, we will replace the `apiVersion` and `kind` fields in the
hncconfig spec with `group` and `resource` to use restmapper lookup to
see if the type exists and remove version in the spec.

Still keep the object reconcilers and in-memory forest to use GVK. In
HNCConfig reconciler and validator, always get the mapping from GR to
GVK first so it can check forest or create object reconcilers with GVK.

Add `version` field in the HNCConfig status to show users which version
is picked. Add `TypeNotFound` condition for HNCConfig when the type is
not found, which was previously reported with
`ObjectReconcilerCreationFailed`. `ObjectReconcilerCreationFailed` will
only be for errors happened when creating the reconciler.

Hack the CRD conversion with string manipulation from `apiVersion` to
`group` and `kind` to `resource`.

Tested by manual testing on GKE cluster, `make test` and `make
test-conversion`. `make test-e2e` passed eventually (had some failure
when running all tests and then `FIt` the failed tests and they all
passed).