Fix Prometheus metrics
We were using the global Prometheus registry, but controller-runtime
(now?) uses its own, separate registry. This change uses
controller-runtime's registry instead.

This change also uses the most recent OpenCensus Prometheus exporter and
removes a bunch of dead code, including direct access of the Prometheus
golang client. As a result, there are a bunch of changes in the vendor
directory.

Tested: directly inspected the /metrics endpoint on the HNC pod via
wget. Without this change, none of the custom HNC metrics are included;
with this change, they are.