Change HNC Config "types" to "resources"
We used the term "type" to represent GVKs, but now that we no longer
care about versions and have changed from kinds to resources, the term
"resources" now seems more appropriate. This commit makes that change.

This change also makes the `group` field optional for builtin types,
such as secrets. Finally, it changes the `hns config set-type` command
to `hns config set-resource` and changes the flags so that the ordered
arg is the resource name, and the propagation mode has an explicit
`--mode` flag (the opposite of what it used to be) since this reads more
naturally.

Also silenced two particularly noisy reconciler messages.

Tested: updated the HNCConfig conversion tests and they all passed.
Updated e2e tests and they passed too.  Played a bit with the HNC config
by hand and confirmed that everything looks good.