Reenable test for #1155
I wasn't able to get the original problem to reproduce itself, but I
added a log message and improve the test so that we can catch #1155 if
it ever reoccurs.

Tested: the newly enabled test passes with HNC as-is; if I break HNC to
manually re-enable allowCascadingDeletion, then the test fails as
expected. I also manually verified that the new log message is displayed
as expected.