Don't block object deletion in dying namespaces
See issue #1214. Our e2e tests were occasionally dying when a namespace
got messed up while propagated objects were in it; I *think* this is
because they got into an ActivityHalted state (so propagated objects
weren't cleaned up) but then the object validator would block deletion.
This caused the tests to hang indefinitely, and the objects couldn't
(easily) be removed without disbling the validator.

This change lets the validators know when the namespace is being
deleted, and unblocks deletion of all propagated objects.

Tested: a new e2e test hangs indefinitely without the fix and completes
successfully with it.