Make 'group' field in config.status.resources omitempty
We have the 'group' field in the config.spec.resources omitempty for K8s
core resources. This PR makes the field in status matches what in the
spec. For K8s core resources, the group will be omitted.

Tested manually by configuring 'secrets', which is a K8s core resource
with empty group and the group field is omitted in both spec and status.