Fix E2E tests and namespace deletion validation
When deleting a parent namespace, we should allow it if none of its
children are sub-namespace.

In the E2E test, I moved the webhook deletion out of AfterEach, so that
if the user doesn't have HNC_REPAIR path set, it won't delete the
webhook.

Since we changed the deletion validator, one of the tests for issue #716
does not apply anymore. I deleted it in this PR.

Tested: make test-e2e