Exclude kube-node-lease from HNC
We shouldn't allow subnamespaces or nontrivial hierarchies in
kube-node-lease, just like it's not allowed in kube-system and a few
others.

Tested: without this change, can create a subns in kube-node-lease; with
this change, I get the appropriate error message. Not adding an
automated test for this since it would just duplicate the contents of
the default_config.go file. All existing e2e tests pass.