Support unpropagated annotations
External systems like ConfigSync use labels and annotations to
understand which objects they should control, just like HNC does. If HNC
creates propagated objects with those labels and annotations, the
external system thinks it owns the objects and can try to modify them
(e.g. delete them - it didn't think they should be created in the first
place so it doesn't know why HNC is creating them).

This commit makes two changes:

* It sets the standard `app.kubernetes.io/managed-by` label on all
  propagated objects. This is very similar to the
  `hnc.x-k8s.io/managed-by` annotation we respect on the namespaces; we
  should probably switch to using this in a future change (Config Sync
  currently sets both).
* It adds an UnpropagatedAnnotations field to the HNCConfiguration
  object. Any annotation listed in this field will not be propagated.
  Config Sync doesn't currently need any unpropagated *labels* but these
  could easily be added if necessary.

Both of these changes are required to get HNC to play well with Config
Sync, with the following three annotations marked as unpropagated (I
believe only the fist one is strictly necessary):

  - configmanagement.gke.io/managed
  - configmanagement.gke.io/declared-config
  - configmanagement.gke.io/token

Tested: since this feature is designed to work as part of an end-to-end
system, most of my testing was end-to-end (apart from the added tests).
Here's what I tested with Config Sync and a simple unstructured repo
with HNC `HierarchyConfiguration` objects checked in:

* I verified that `app.kubernetes.io/managed-by` label is set correctly
  on a few propagated objects.
* In the log, I verified that the HNC webhooks were constantly rejecting
  "delete" requests from Config Sync.
* I added the three annotations shown above to the HNC Config object,
  and verified in the log that the object reconcilers were asked to full
  resync all objects, and that all source objects were indeed resynced,
  and that the requests from Config Sync stopped immediately after the
  objects were updated.
* I verified by hand that none of the three annotations appeared on a
  couple of propagated objects.
* I removed the three annotations from the HNCConfiguration and verified
  that the annotations reappeared on the objects and that Config Sync
  was once again fighting HNC. I then re-added them a final time and
  everything went back to normal.

Any further end-to-end testing will be done on the Hierarchy Controller
side, not in OSS HNC.