Tool: map abstract to hierarchical namespaces
This is an experimental tool to inspect a directory structure containing
abstract namespaces (from GKE Config Sync or Anthos Config Management)
and add HNC objects so the directory can be used with `kubectl apply -f`
and recreate the directory structure using hierarchical namespaces.

Tested: ran the tool as described in the readme on the examples/simple-1
directory structure and verified that, when applied to a cluster, all
the expected hierarchical relationships were created.