Set unpropagated annotations via command line args
The list of unpropagated annotations should very rarely be changed, and
controlling them via the HNCConfiguration object has a bunch of problems
(e.g. merging fields set by an external operator, like the Config Sync
operator, with user-specified fields). As a result, I think it makes
more sense to specify these as command-line params instead. This also
simplifies a bunch of logic.

Also documented all potentially useful command-line args.

Tested: verified that the annoations are removed if specified in the
command line args in the Deployment and that HNC plays well with Config
Sync as a result. Migrated the HNCConfiguration integ test to the object
reconciler instead.