Fix Docker-in-Docker and image name
This change uses the krte (KIND RunTime Environment) image as a base,
since it knows how to run Docker-in-Docker even when the Prow node
doesn't have Docker installed on it already (e.g. because it uses
containerd, I think).

Also sets the HNC_REGISTRY env var explicitly so it doesn't pick up the
default from gcloud in the image.

Tested: Without this change, HNC periodics are failing; with a slightly
earlier version of this change, everything passes. Now that the
periodcis are only running every 48h, it's harder to test this before
submitting, but the postsubmits should pick up this change and run it
(I've already pushed the updated image) so if there's a problem we'll
find out quickly.