Attempt to fix HNC postsubmits
Our first run of the HNC postsubmits failed, most likely because the
workdir is modified by Prow (based on the messages I'm seeing). This
change updates the Dockerfile and test script to _not_ assume a specific
workdir, and also allows Prow to set up the repo when appropriate.

Tested: script runs locally but there's no good way for me to test on
Prow. As soon as this is merged, the postsubmits themselves will tell me
if I got this right or not.