Prepare to ditch custom Prow image
It turns out that Prow replaces the entire $GOPATH in postsubmits (and,
optionally, in periodics) so installing Kind in the Dockerfile isn't
terribly useful. As a result, this change strips down the Docker image
(which I've already pushed) to basically be a shell that does nothing
but contain the test script. A future change to the test-infra config
will bypass the custom image completely, at which point we can remove it
entirely.

In preparation for this, I've also hardlinked the test script to the
parent directory, since I'll also get rid of the prow-e2e directory as
well and merge all useful information back into the parent.

The script itself has been modified to download Kind itself. The code to
clone the repo will also shortly be removed once the periodics have been
configured to get Prow to do it for us. Finally, the Makefile has been
modified to remove the unused build targets and invoke the test script
as it will (shortly) be invoked by Prow.

Tested: I pushed the stripped-down Docker image to the registry, and
then (before updating the makefile to point to the base krte image)
verified that the tests still ran. I also tested the script using the
base krte image (e.g. what's being checked in) to make sure it will work
after I update test-infra.