Update cert-rotator to 0.2.0; tweak test targets
This change uses an "official" cert-rotator release for the first time,
and includes support for v1 of all K8s APIs.

It also introduces the "test-smoke" target that only runs the quickstart
tests, to allow for faster local development. This allows developers to
rely more heavily on the postsubmit and periodic tests we worked so hard
to set up for HNC.

Finally, it clarifies some of the warnings around HNC_REPAIR.

Tested: make test-smoke passes. Tried running 'make test-e2e' with and
without HNC_REPAIR set to verify that the warnings look correct in both
cases.