Support cluster-admin propagation and update docs
Rather than continuing to play whack-a-mole with the list of verbs HNC
can propagate, this change allows it to perform all verbs on all
resources. This is equivalent to `cluster-admin` so I've updated the
docs accordingly (see also issue #1311).

I also noticed that the docs referred to K8s v1.15, which is no longer
supported, so I updated them to v1.16.

Finally, this change adds the `HNC_FOCUS` makefile var, allowing you to
say something like:

```
HNC_FOCUS=772 make test
```

which only runs the e2e tests with "772" in the title.

Tested: All quickstart e2e tests pass. Updated the test for issue #772
and verified that it failed without the other changes in this commit,
and passed with them.