Allow HNC to propagate changes to CRD.
Currently, HNC are clearing out most of the metadata when doing an update request, which works fine for most types, but failed for CRDs. This PR allows HNC to propagate changes to CRD by setting 'ResourceVersion' explicitly when performing an update.
Tested: before this change, first create an crd in the parent namespace, and then update whatever field of the crd, wait a second and run the following command:
kubectl hns describe child

Result:

Events from the objects in namespace child
Last Seen Reason             Object                     Message
0s        CannotUpdateObject EETest:child/eetest-sample Could not write from source namespace parent: eetests.e2e.hnc.x-k8s.io eetest-sample is invalid: metadata.resourceVersion: Invalid value: 0x0: must be specified for an update.

After this change, you should see it propagates normally and the inherited crd's field is updated.